+++ RELEASENOTES.md
@@ -66,6 +66,9 @@
 *   PlayerNotificationManager:
     *   Add `PendingIntent.FLAG_IMMUTABLE` flag to BroadcastReceiver to support
         Android 12.
+    *   Add `setUseFastForwardActionInCompactView(boolean)` and
+        `setUseRewindActionInCompactView(boolean)` to make it possible to show
+        seek actions in compact view mode.
 *   DRM:
     *   Don't restore offline keys before releasing them. In OEMCrypto v16+ keys
         must be released without restoring them first.